requirements.txt
python-dotenv
requests
openai
pydantic-ai
crawl4ai
websockets
pyee
pyppeteer
supabase 
logfire
streamlit